+++ playground/components/StarlinkDemo.vue
@@ -18,8 +18,16 @@
 
     <UCard class="p-4">
       <div class="mb-4">
-        <label for="limit" class="mr-2">Limit:</label>
-        <input v-model.number="limit" class="w-10" type="number" min="1">
+        <label
+          for="limit"
+          class="mr-2"
+        >Limit:</label>
+        <input
+          v-model.number="limit"
+          class="w-10"
+          type="number"
+          min="1"
+        >
       </div>
 
       <div>
@@ -35,7 +43,6 @@
 </template>
 
 <script lang="ts" setup>
-// @ts-ignore
 import queryLaunches from '~/queries/launches.gql'
 
 const queryShips = gql`
@@ -55,7 +62,7 @@ const { load, onError, refetch, result: launchResult } = useLazyQuery(queryLaunc
   fetchPolicy: 'no-cache'
 })
 watch(launchResult, v => (data.value = v))
-// eslint-disable-next-line no-console
+
 onError(e => console.error(e))
 
 const getLaunches = () => !launchResult.value ? load() : refetch()