+++ playground/components/TodosDemo.vue
@@ -10,11 +10,17 @@
           Load Todos
         </UButton>
 
-        <UButton :disabled="!subscribe" @click="createTodo">
+        <UButton
+          :disabled="!subscribe"
+          @click="createTodo"
+        >
           Create Todo
         </UButton>
 
-        <UButton :disabled="subscribe" @click="todoAdded">
+        <UButton
+          :disabled="subscribe"
+          @click="todoAdded"
+        >
           Subscribe
         </UButton>
       </div>
@@ -41,7 +47,7 @@ const { data, refresh } = await useAsyncQuery<TodoEntry[]>(gqlTodos, null, 'todo
 
 const { mutate: todoMutation } = useMutation(gqlCreateTodo, { clientId: 'todos' })
 
-function createTodo () {
+function createTodo() {
   todoMutation({
     todo: {
       text: 'Random ' + Math.floor(Math.random() * 100)
@@ -51,17 +57,16 @@ function createTodo () {
 
 const subscribe = ref(false)
 
-function todoAdded () {
+function todoAdded() {
   subscribe.value = true
 
   const { onResult, onError } = useSubscription(gqlTodoAdded, null, { clientId: 'todos' })
 
   onResult((r) => {
-    data.value = r.data as any
+    data.value = r.data as TodoEntry[]
   })
 
   onError((e) => {
-    // eslint-disable-next-line no-console
     console.log(e)
   })
 }