+++ src/runtime/composables.ts
@@ -6,8 +6,9 @@ import type { RestartableClient } from './ws'
 import { ref, unref, isRef, reactive, useCookie, useNuxtApp, useAsyncData } from '#imports'
 import { NuxtApollo } from '#apollo'
 import type { ApolloClientKeys } from '#apollo'
+import type { ClientConfig } from '../types'
 
-type PickFrom<T, K extends Array<string>> = T extends Array<any> ? T : T extends Record<string, any> ? keyof T extends K[number] ? T : K[number] extends never ? T : Pick<T, K[number]> : T
+type PickFrom<T, K extends Array<string>> = T extends Array<unknown> ? T : T extends Record<string, unknown> ? keyof T extends K[number] ? T : K[number] extends never ? T : Pick<T, K[number]> : T
 type KeysOf<T> = Array<T extends T ? keyof T extends string ? keyof T : never : never>
 
 type TQuery<T> = QueryOptions<OperationVariables, T>['query']
@@ -37,7 +38,7 @@ type TAsyncQuery<T> = {
   context?: DefaultContext
   /**
    * If `true`, this overrides the default fetchPolicy for the Apollo Client to `cache-first`.
-   * */
+   */
   cache?: boolean
 }
 
@@ -47,13 +48,15 @@ type TAsyncQuery<T> = {
  * @param opts An object containing the query, variables, clientId, context, and cache options.
  * @param options Customize the underlying `useAsyncData` composable.
  */
-export function useAsyncQuery <
+export function useAsyncQuery<
   T,
   DataT = T,
   PickKeys extends KeysOf<DataT> = KeysOf<DataT>,
   DefaultT = null,
   NuxtErrorDataT = unknown
-> (opts: TAsyncQuery<T>, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+>(opts: TAsyncQuery<T>, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+
+/* eslint-disable jsdoc/check-param-names */
 
 /**
  * `useAsyncQuery` resolves the GraphQL query asynchronously in a SSR-friendly composable.
@@ -64,15 +67,15 @@ export function useAsyncQuery <
  * @param context The context object that's passed along your link chain.
  * @param options Customize the underlying `useAsyncData` composable.
  */
-export function useAsyncQuery <
+export function useAsyncQuery<
   T,
   DataT = T,
   PickKeys extends KeysOf<DataT> = KeysOf<DataT>,
   DefaultT = null,
   NuxtErrorDataT = unknown
-> (query: TQuery<T>, variables?: TVariables<T>, clientId?: ApolloClientKeys, context?: DefaultContext, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+>(query: TQuery<T>, variables?: TVariables<T>, clientId?: ApolloClientKeys, context?: DefaultContext, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
 
-export function useAsyncQuery <T> (...args: any[]) {
+export function useAsyncQuery<T>(...args: unknown[]) {
   const { key, fn, options } = prep<T>(...args)
   return useAsyncData<T>(key, fn, options)
 }
@@ -83,13 +86,13 @@ export function useAsyncQuery <T> (...args: any[]) {
  * @param opts An object containing the query, variables, clientId, context, and cache options.
  * @param options Customize the underlying `useAsyncData` composable.
  */
-export function useLazyAsyncQuery <
+export function useLazyAsyncQuery<
   T,
   DataT = T,
   PickKeys extends KeysOf<DataT> = KeysOf<DataT>,
   DefaultT = null,
   NuxtErrorDataT = unknown
-> (opts: TAsyncQuery<T>, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+>(opts: TAsyncQuery<T>, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
 
 /**
  * `useLazyAsyncQuery` resolves the GraphQL query after loading the route, instead of blocking client-side navigation.
@@ -100,19 +103,21 @@ export function useLazyAsyncQuery <
  * @param context The context object that's passed along your link chain.
  * @param options Customize the underlying `useAsyncData` composable.
  */
-export function useLazyAsyncQuery <
+export function useLazyAsyncQuery<
   T,
   DataT = T,
   PickKeys extends KeysOf<DataT> = KeysOf<DataT>,
   DefaultT = null,
   NuxtErrorDataT = unknown
-> (query: TQuery<T>, variables?: TVariables<T>, clientId?: string, context?: DefaultContext, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+>(query: TQuery<T>, variables?: TVariables<T>, clientId?: string, context?: DefaultContext, options?: AsyncDataOptions<T, DataT, PickKeys, DefaultT>): AsyncData<PickFrom<DataT, PickKeys> | DefaultT, (NuxtErrorDataT extends Error | NuxtError ? NuxtErrorDataT : NuxtError<NuxtErrorDataT>) | null>
+/* eslint-enable jsdoc/check-param-names */
 
-export function useLazyAsyncQuery <T> (...args: any) {
+export function useLazyAsyncQuery<T>(...args: unknown[]) {
   const { key, fn, options } = prep<T>(...args)
   return useAsyncData<T>(key, fn, { ...options, lazy: true })
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 const prep = <T> (...args: any[]) => {
   const { clients } = useApollo()
 
@@ -125,15 +130,16 @@ const prep = <T> (...args: any[]) => {
 
   let options: AsyncDataOptions<T, T, KeysOf<T>, null> = {}
 
-  if ((typeof args?.[0] === 'object' && 'query' in args[0])) {
-    query = args?.[0]?.query
-    variables = args?.[0]?.variables
+  if ((typeof args?.[0] === 'object' && args[0] !== null && 'query' in args[0])) {
+    const firstArg = args?.[0] as TAsyncQuery<T>
+    query = firstArg.query
+    variables = firstArg.variables
 
-    cache = args?.[0]?.cache
-    context = args?.[0]?.context
-    clientId = args?.[0]?.clientId
+    cache = firstArg.cache!
+    context = firstArg.context!
+    clientId = firstArg.clientId
 
-    if (typeof args?.[1] === 'object') {
+    if (typeof args?.[1] === 'object' && args?.[1] !== null) {
       options = args?.[1]
     }
   } else {
@@ -148,12 +154,16 @@ const prep = <T> (...args: any[]) => {
     }
   }
 
-  if (!query) { throw new Error('@nuxtjs/apollo: no query provided') }
+  if (!query) {
+    throw new Error('@nuxtjs/apollo: no query provided')
+  }
 
   if (!clientId || !clients?.[clientId]) {
     clientId = (clients?.default ? 'default' : Object.keys(clients!)?.[0]) as ApolloClientKeys
 
-    if (!clientId) { throw new Error('@nuxtjs/apollo: no client found') }
+    if (!clientId) {
+      throw new Error('@nuxtjs/apollo: no client found')
+    }
   }
 
   if (variables) {
@@ -175,11 +185,11 @@ const prep = <T> (...args: any[]) => {
   return { key, query, clientId, variables, fn, options }
 }
 
-export function useApollo (): {
+export function useApollo(): {
   /**
    * Access the configured apollo clients.
    */
-  clients: Record<ApolloClientKeys, ApolloClient<any>> | undefined
+  clients: Record<ApolloClientKeys, ApolloClient<unknown>> | undefined
   /**
    * Retrieve the auth token for the specified client. Adheres to the `apollo:auth` hook.
    *
@@ -193,7 +203,7 @@ export function useApollo (): {
    * @param {string} token The token to be applied.
    * @param {string} client - Name of the Apollo client. Defaults to `default`.
    * @param {boolean} skipResetStore - If `false`, Resets your entire store by clearing out your cache and then re-executing all of your active queries.
-   * */
+   */
   onLogin: (token?: string, client?: ApolloClientKeys, skipResetStore?: boolean) => Promise<void>
 
   /**
@@ -201,13 +211,13 @@ export function useApollo (): {
    *
    * @param {string} client - Name of the Apollo client. Defaults to `default`.
    * @param {boolean} skipResetStore - If `false`, Resets your entire store by clearing out your cache and then re-executing all of your active queries.
-   * */
+   */
   onLogout: (client?: ApolloClientKeys, skipResetStore?: boolean) => Promise<void>
 }
 
-export function useApollo () {
+export function useApollo() {
   const nuxtApp = useNuxtApp() as NuxtApp & {
-    _apolloClients?: Record<ApolloClientKeys, ApolloClient<any>>
+    _apolloClients?: Record<ApolloClientKeys, ApolloClient<unknown>>
     _apolloWsClients?: Record<ApolloClientKeys, RestartableClient>
   }
 
@@ -216,39 +226,48 @@ export function useApollo () {
 
     const conf = NuxtApollo?.clients?.[client]
 
-    if (!conf) { return }
+    if (!conf) {
+      return
+    }
 
     const token = ref<string | null>(null)
     await (nuxtApp as ReturnType<typeof useNuxtApp>).callHook('apollo:auth', { token, client })
 
-    if (token.value) { return token.value }
+    if (token.value) {
+      return token.value
+    }
 
     const tokenName = conf.tokenName!
 
     return conf?.tokenStorage === 'cookie'
       ? nuxtApp.runWithContext(() => useCookie(tokenName).value)
-      : (process.client && localStorage.getItem(tokenName)) || null
+      : (import.meta.client && localStorage.getItem(tokenName)) || null
   }
-  type TAuthUpdate = {token?: string, client?: ApolloClientKeys, mode: 'login' | 'logout', skipResetStore?: boolean}
+  type TAuthUpdate = { token?: string, client?: ApolloClientKeys, mode: 'login' | 'logout', skipResetStore?: boolean }
   const updateAuth = async ({ token, client, mode, skipResetStore }: TAuthUpdate) => {
     client = client || 'default'
 
     const conf = NuxtApollo?.clients?.[client]
 
-    if (!conf) { return }
+    if (!conf) {
+      return
+    }
 
     const tokenName = client && conf.tokenName!
 
     if (conf?.tokenStorage === 'cookie') {
-      const cookieOpts = (client && conf?.cookieAttributes) || NuxtApollo?.cookieAttributes
+      const cookieOpts = ((client && conf?.cookieAttributes) || NuxtApollo?.cookieAttributes) as ClientConfig['cookieAttributes'] & {
+        readonly?: false
+      }
 
-      // @ts-ignore
       const cookie = useCookie(tokenName, cookieOpts)
 
-      if (!cookie.value && mode === 'logout') { return }
+      if (!cookie.value && mode === 'logout') {
+        return
+      }
 
       cookie.value = (mode === 'login' && token) || null
-    } else if (process.client && conf?.tokenStorage === 'localStorage') {
+    } else if (import.meta.client && conf?.tokenStorage === 'localStorage') {
       if (mode === 'login' && token) {
         localStorage.setItem(tokenName, token)
       } else if (mode === 'logout') {
@@ -256,12 +275,15 @@ export function useApollo () {
       }
     }
 
-    if (nuxtApp?._apolloWsClients?.[client]) { nuxtApp._apolloWsClients[client].restart() }
+    if (nuxtApp?._apolloWsClients?.[client]) {
+      nuxtApp._apolloWsClients[client]?.restart()
+    }
 
-    if (skipResetStore) { return }
+    if (skipResetStore) {
+      return
+    }
 
-    // eslint-disable-next-line no-console
-    await nuxtApp?._apolloClients?.[client].resetStore().catch(e => console.log('%cError on cache reset', 'color: orange;', e.message))
+    await nuxtApp?._apolloClients?.[client]?.resetStore().catch(e => console.log('%cError on cache reset', 'color: orange;', e.message))
   }
 
   return {